CONFIRMED WORKING; migrated CharAssigner Flash app into a class, kludgily
