+++ oot3dhdtextgenerator/apps/char_assigner/routes/__init__.py
