--- oot3dhdtextgenerator/apps/char_assigner/views/__init__.py
@@ -1,18 +1,0 @@
-#  Copyright 2020-2023 Karl T Debiec. All rights reserved. This software may be modified
-#  and distributed under the terms of the BSD license. See the LICENSE file for details.
-from oot3dhdtextgenerator.apps.char_assigner.views.books import (
-    create,
-    delete,
-    edit,
-    read,
-    update,
-)
-from oot3dhdtextgenerator.apps.char_assigner.views.index import index
-
-__all__ = [
-    "create",
-    "delete",
-    "edit",
-    "read",
-    "update",
-]
