+++ oot3dhdtextgenerator/cli/char_assigner_cli.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 from argparse import ArgumentParser
 from typing import Any
 
-from oot3dhdtextgenerator.apps.char_assigner import app
+from oot3dhdtextgenerator.apps.char_assigner import CharAssigner
 from oot3dhdtextgenerator.common import (
     CommandLineInterface,
     get_arg_groups_by_name,
@@ -42,7 +42,7 @@ class CharAssignerCli(CommandLineInterface):
             type=int_arg(min_value=10, max_value=9933),
             default=1000,
             help="number of characters included in model, starting from the most "
-            "common and ending with the least common (default: %(default)d, max: 9933)",
+                 "common and ending with the least common (default: %(default)d, max: 9933)",
         )
         arg_groups["input arguments"].add_argument(
             "--assignment-file",
@@ -76,7 +76,8 @@ class CharAssignerCli(CommandLineInterface):
     @classmethod
     def main_internal(cls, **kwargs: Any) -> None:
         """Execute with provided keyword arguments."""
-        app.run(port=5001)
+        char_assigner = CharAssigner(**kwargs)
+        char_assigner.run(port=5001)
 
 
 if __name__ == "__main__":
